+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## Next_Ver - 2021/12/xx
+
+- Add 'Enable Keyboard' section with 'Launch Settings' button. https://github.com/FredHappyface/Android.EweSticker/issues/31
+- Show back button in navbar by default. https://github.com/FredHappyface/Android.EweSticker/issues/32
+
 ## 20211118 - 2021/11/18
 
 - Attempt to resolve issue reported where the app crashed when importing stickers (suspected cause